refactor(comp:table): extract expanded row render
